+++ test/test_name_da.rb
@@ -23,18 +23,18 @@ class TestFakerNameDA < Test::Unit::TestCase
   end
 
   def test_any_name
-    assert @tester.any_name.count(' ') < 3
+    assert_less_than @tester.any_name.count(' '), 3
   end
 
   def test_male_name
     regexp = /#{@tester::MALE_FIRST_NAMES.join('|')}/
-    assert @tester.male_name.count(' ') < 3
+    assert_less_than @tester.male_name.count(' '), 3
     assert_match(regexp, @tester.male_name)
   end
 
   def test_female_name
     regexp = /#{@tester::FEMALE_FIRST_NAMES.join('|')}/
-    assert @tester.female_name.count(' ') < 3
+    assert_less_than @tester.female_name.count(' '), 3
     assert_match(regexp, @tester.female_name)
   end